[Update] モノビヘイビアーを継承したシングルトンクラスの移動。
